+++ app-modules/service-management/src/Filament/Resources/IncidentUpdateResource.php
@@ -118,10 +118,6 @@ class IncidentUpdateResource extends Resource
             ]);
     }
 
-    public static function getRelations(): array
-    {
-        return [];
-    }
 
     public static function getPages(): array
     {
