+++ app-modules/service-management/src/Filament/Resources/IncidentUpdateResource/Pages/EditIncidentUpdate.php
@@ -38,7 +38,8 @@ namespace AidingApp\ServiceManagement\Filament\Resources\IncidentUpdateResource\
 
 use AidingApp\ServiceManagement\Filament\Resources\IncidentUpdateResource;
 use App\Concerns\EditPageRedirection;
-use Filament\Actions;
+use Filament\Actions\DeleteAction;
+use Filament\Actions\ViewAction;
 use Filament\Resources\Pages\EditRecord;
 
 class EditIncidentUpdate extends EditRecord
@@ -50,8 +51,8 @@ class EditIncidentUpdate extends EditRecord
     protected function getHeaderActions(): array
     {
         return [
-            Actions\ViewAction::make(),
-            Actions\DeleteAction::make(),
+            ViewAction::make(),
+            DeleteAction::make(),
         ];
     }
 }
