+++ app-modules/service-management/src/Filament/Resources/IncidentUpdateResource/Pages/ListIncidentUpdates.php
@@ -37,7 +37,7 @@
 namespace AidingApp\ServiceManagement\Filament\Resources\IncidentUpdateResource\Pages;
 
 use AidingApp\ServiceManagement\Filament\Resources\IncidentUpdateResource;
-use Filament\Actions;
+use Filament\Actions\CreateAction;
 use Filament\Resources\Pages\ListRecords;
 
 class ListIncidentUpdates extends ListRecords
@@ -47,7 +47,7 @@ class ListIncidentUpdates extends ListRecords
     protected function getHeaderActions(): array
     {
         return [
-            Actions\CreateAction::make(),
+            CreateAction::make(),
         ];
     }
 }
